Add Javadoc to Fixed class
Other changes
- Remove unnecessary type casting

This resolves #249

+++ autoparams/src/main/java/org/javaunit/autoparams/generator/DoubleGenerator.java
@@ -25,7 +25,7 @@ final class DoubleGenerator implements ObjectGenerator {
             ArgumentQuery argumentQuery = (ArgumentQuery) query;
             Min annotation = argumentQuery.getParameter().getAnnotation(Min.class);
             if (annotation != null) {
-                return (double) Math.max(Double.MIN_VALUE, annotation.value());
+                return Math.max(Double.MIN_VALUE, annotation.value());
             }
         }
